# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiResponse_DeploymentResponse_": {
        "properties": {
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DeploymentResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ErrorDetail"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "title": "ApiResponse[DeploymentResponse]",
        "type": "object"
      },
      "ApiResponse_DiagnosisResponse_": {
        "properties": {
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DiagnosisResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ErrorDetail"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "title": "ApiResponse[DiagnosisResponse]",
        "type": "object"
      },
...
      "ApiResponse_NoneType_": {
        "properties": {
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          },
          "data": {
            "title": "Data",
            "type": "null"
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ErrorDetail"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "title": "ApiResponse[NoneType]",
        "type": "object"
      },
      "ApiResponse_Page_DeploymentResponse__": {
        "properties": {
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Page_DeploymentResponse_"
              },
              {
                "type": "null"
              }
            ]
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ErrorDetail"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "title": "ApiResponse[Page[DeploymentResponse]]",
        "type": "object"
      },
...
      "ApiResponse_dict_str__Any__": {
        "properties": {
          "code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code"
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ErrorDetail"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          }
        },
        "title": "ApiResponse[dict[str, Any]]",
        "type": "object"
      },
...
      "CreateDeploymentRequest": {
        "description": "배포 요청을 직접 만든다. 푸시 웹훅이 못 하는 첫 배포·CLI 업로드·재배포·롤백·재시작·삭제.",
        "properties": {
          "sourceDeploymentId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "REDEPLOY·ROLLBACK 에서 필수. 같은 서비스의 이전 배포 id. ROLLBACK 은 SUCCEEDED 여야 한다. MANUAL·RESTART·REMOVE 에서는 보내지 않는다.",
            "examples": [
              12
            ],
            "title": "Sourcedeploymentid"
          },
          "sourceSha": {
            "anyOf": [
              {
                "maxLength": 64,
                "minLength": 7,
                "pattern": "^[0-9a-fA-F]+$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "MANUAL 에서만 쓴다. 없으면 서비스 브랜치의 최신 커밋을 서버가 조회한다.",
            "examples": [
              "9f2c1ab"
            ],
            "title": "Sourcesha"
          },
          "triggerType": {
            "description": "MANUAL: 브랜치 최신 커밋(또는 sourceSha)을 빌드해 배포한다. CLI: `POST /services/{serviceId}/uploads` 로 올린 소스(uploadId)를 빌드해 배포한다. REDEPLOY: sourceDeploymentId 의 커밋을 다시 빌드해 배포한다(CLI 업로드로 만든 배포는 소스가 남아 있지 않아 다시 빌드할 수 없다). ROLLBACK: 성공했던 sourceDeploymentId 가 만든 이미지를 빌드 없이 그대로 배포한다. RESTART: 지금 떠 있는(마지막으로 성공한) 배포의 이미지를 빌드 없이 다시 띄워 Pod 를 새로 시작한다. REMOVE: 지금 떠 있는 배포를 클러스터에서 내린다(서비스 정의는 남는다). 다시 배포하려면 MANUAL·CLI·REDEPLOY·ROLLBACK 을 쓴다.",
            "enum": [
              "MANUAL",
              "CLI",
              "REDEPLOY",
              "ROLLBACK",
              "RESTART",
              "REMOVE"
            ],
            "examples": [
              "MANUAL"
            ],
            "title": "Triggertype",
            "type": "string"
          },
          "uploadId": {
            "anyOf": [
              {
                "maxLength": 64,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "CLI 에서 필수. 같은 서비스에 올린 업로드의 ID. 업로드 하나는 배포 요청 하나에만 쓸 수 있다. 다른 트리거에서는 보내지 않는다.",
            "examples": [
              "Zq3h8mP0xYkq2oVd1sT6uXn4wJc9bLrE5aGfHiK7yMA"
            ],
            "title": "Uploadid"
          }
        },
        "required": [
          "triggerType"
        ],
        "title": "CreateDeploymentRequest",
        "type": "object"
      },
...
      "DeploymentResponse": {
        "description": "배포 요청 1건. 상태 이름: QUEUED=Initializing, SUCCEEDED=Active.",
        "properties": {
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "failureCode": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FailureCode"
              },
              {
                "type": "null"
              }
            ],
            "description": "FAILED 일 때만 있다. 원인 구분은 상태가 아니라 이 코드로 한다."
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "isActive": {
            "description": "QUEUED·BUILDING·DEPLOYING 이면 진행 중이다.",
            "title": "Isactive",
            "type": "boolean"
          },
          "requestedBy": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "요청한 사용자 id. 푸시 웹훅이 만든 요청은 없다.",
            "title": "Requestedby"
          },
          "serviceId": {
            "title": "Serviceid",
            "type": "integer"
          },
          "sourceCommitMessage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sourcecommitmessage"
          },
          "sourceDeploymentId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "재배포·롤백·재시작·삭제가 따라간 원본 배포 id. 롤백·재시작은 이 배포가 만든 이미지를 쓰고, 삭제는 이 배포를 내린다.",
            "title": "Sourcedeploymentid"
          },
          "sourceSha": {
            "title": "Sourcesha",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/DeploymentStatus"
          },
          "triggerType": {
            "$ref": "#/components/schemas/DeploymentTrigger"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "id",
          "serviceId",
          "status",
          "sourceSha",
          "triggerType",
          "isActive",
          "createdAt",
          "updatedAt"
        ],
        "title": "DeploymentResponse",
        "type": "object"
      },
...
      "DeploymentStatus": {
        "enum": [
          "QUEUED",
          "BUILDING",
          "DEPLOYING",
          "SUCCEEDED",
          "FAILED",
          "ROLLED_BACK",
          "MANUAL_INTERVENTION",
          "SUPERSEDED"
        ],
        "title": "DeploymentStatus",
        "type": "string"
      },
...
      "DeploymentTrigger": {
        "enum": [
          "MANUAL",
          "PUSH",
          "CLI",
          "REDEPLOY",
          "ROLLBACK",
          "RESTART",
          "REMOVE"
        ],
        "title": "DeploymentTrigger",
        "type": "string"
      },
...
      "ErrorDetail": {
        "description": "필드별 검증 실패 사유. field 는 camelCase 필드 경로다.",
        "properties": {
          "field": {
            "title": "Field",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "field",
          "reason"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
...
      "FailureCode": {
        "enum": [
          "SOURCE_NOT_ACCESSIBLE",
          "SOURCE_REF_NOT_FOUND",
          "SOURCE_TOO_LARGE",
          "SOURCE_INVALID",
          "BUILD_CONFIG_REQUIRED",
          "BUILD_FAILED",
          "BUILD_TIMED_OUT",
          "BUILD_INFRA_ERROR",
          "DEPLOY_FAILED",
          "DEPLOY_TIMED_OUT",
          "DEPLOY_INFRA_ERROR"
        ],
        "title": "FailureCode",
        "type": "string"
      },
...
      "Page_DeploymentResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DeploymentResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "size"
        ],
        "title": "Page[DeploymentResponse]",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "in": "cookie",
        "name": "anydeploy_session",
        "type": "apiKey"
      },
      "HTTPBearer": {
        "description": "CLI 용 세션 토큰",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "모든 JSON 응답은 `ApiResponse` 봉투(`success`·`code`·`message`·`data`)로 감싼다. 필드는 camelCase 다. 인증은 쿠키(웹) 또는 `Authorization: Bearer`(CLI).",
    "title": "AnyDeploy Control API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/services/{service_id}/deployments": {
      "get": {
        "operationId": "search_deployment_requests_api_v1_services__service_id__deployments_get",
        "parameters": [
          {
            "in": "path",
            "name": "service_id",
            "required": true,
            "schema": {
              "title": "Service Id",
              "type": "integer"
            }
          },
          {
            "description": "0부터 시작",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 0,
              "description": "0부터 시작",
              "minimum": 0,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_Page_DeploymentResponse__"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "로그인이 필요하다. 웹훅은 서명이 올바르지 않을 때 (UNAUTHORIZED)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "대상을 찾을 수 없다 (NOT_FOUND · PROJECT_NOT_FOUND · SERVICE_NOT_FOUND · DEPLOYMENT_REQUEST_NOT_FOUND · VARIABLE_NOT_FOUND · DIAGNOSIS_NOT_FOUND · UPLOAD_NOT_FOUND)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "입력이 올바르지 않다 (VALIDATION_ERROR · INVALID_INPUT)"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "summary": "배포 요청 목록 (최신순)",
        "tags": [
          "deployments"
        ]
      },
      "post": {
        "operationId": "create_deployment_request_api_v1_services__service_id__deployments_post",
        "parameters": [
          {
            "in": "path",
            "name": "service_id",
            "required": true,
            "schema": {
              "title": "Service Id",
              "type": "integer"
            }
          },
          {
            "description": "같은 값으로 다시 보내면 새로 만들지 않고 처음 만든 배포 요청을 돌려준다.",
            "in": "header",
            "name": "idempotency-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "같은 값으로 다시 보내면 새로 만들지 않고 처음 만든 배포 요청을 돌려준다.",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDeploymentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_DeploymentResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "로그인이 필요하다. 웹훅은 서명이 올바르지 않을 때 (UNAUTHORIZED)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "대상을 찾을 수 없다 (NOT_FOUND · PROJECT_NOT_FOUND · SERVICE_NOT_FOUND · DEPLOYMENT_REQUEST_NOT_FOUND · VARIABLE_NOT_FOUND · DIAGNOSIS_NOT_FOUND · UPLOAD_NOT_FOUND)"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "현재 상태와 충돌한다. 이미 있는 이름이거나 진행 중인 배포가 있다 (CONFLICT · PROJECT_NAME_CONFLICT · SERVICE_NAME_CONFLICT · VARIABLE_CONFLICT · DEPLOYMENT_IN_PROGRESS · DEPLOYMENT_NOT_FAILED · DIAGNOSIS_IN_PROGRESS · DIAGNOSIS_LOGS_UNAVAILABLE · UPLOAD_UNAVAILABLE)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "입력이 올바르지 않다 (VALIDATION_ERROR · INVALID_INPUT)"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "외부 시스템(GitHub·Loki·Prometheus·진단 에이전트·S3) 호출에 실패했다 (EXTERNAL_ERROR)"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "필요한 설정이 없다 (NOT_CONFIGURED)"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "summary": "배포 요청 생성 (수동·CLI 업로드·재배포·롤백·재시작·삭제)",
        "tags": [
          "deployments"
        ]
      }
    },
...
    "/api/v1/services/{service_id}/deployments/{deployment_id}/diagnose": {
      "post": {
        "description": "실패한 배포의 런타임 로그(와 가능하면 소스)를 에러 진단 에이전트에 보내 원인과 해결책을 받아 저장한다. 모델을 부르는 데 최대 2분 남짓 걸려 **바로 `202` 와 `status=RUNNING` 으로 답하고**, 진단은 서버가 이어서 실행한다. 결과는 `GET .../diagnosis` 를 `status` 가 `SUCCEEDED`·`FAILED` 가 될 때까지 폴링해 받는다. 성공한 진단이 이미 있으면 모델을 다시 부르지 않고 `200` 으로 그 결과를 돌려주며, refresh=true 로 다시 진단한다. 해결책은 제안일 뿐 서버가 실행하지 않고, 배포 요청의 상태도 바꾸지 않는다.",
        "operationId": "diagnose_deployment_request_api_v1_services__service_id__deployments__deployment_id__diagnose_post",
        "parameters": [
          {
            "in": "path",
            "name": "service_id",
            "required": true,
            "schema": {
              "title": "Service Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "deployment_id",
            "required": true,
            "schema": {
              "title": "Deployment Id",
              "type": "integer"
            }
          },
          {
            "description": "true 면 성공한 진단이 있어도 새로 진단한다(모델 비용이 든다).",
            "in": "query",
            "name": "refresh",
            "required": false,
            "schema": {
              "default": false,
              "description": "true 면 성공한 진단이 있어도 새로 진단한다(모델 비용이 든다).",
              "title": "Refresh",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_DiagnosisResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_DiagnosisResponse_"
                }
              }
            },
            "description": "진단을 시작했다(status=RUNNING). GET 으로 결과를 폴링한다"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "로그인이 필요하다. 웹훅은 서명이 올바르지 않을 때 (UNAUTHORIZED)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "대상을 찾을 수 없다 (NOT_FOUND · PROJECT_NOT_FOUND · SERVICE_NOT_FOUND · DEPLOYMENT_REQUEST_NOT_FOUND · VARIABLE_NOT_FOUND · DIAGNOSIS_NOT_FOUND · UPLOAD_NOT_FOUND)"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "현재 상태와 충돌한다. 이미 있는 이름이거나 진행 중인 배포가 있다 (CONFLICT · PROJECT_NAME_CONFLICT · SERVICE_NAME_CONFLICT · VARIABLE_CONFLICT · DEPLOYMENT_IN_PROGRESS · DEPLOYMENT_NOT_FAILED · DIAGNOSIS_IN_PROGRESS · DIAGNOSIS_LOGS_UNAVAILABLE · UPLOAD_UNAVAILABLE)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "입력이 올바르지 않다 (VALIDATION_ERROR · INVALID_INPUT)"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "필요한 설정이 없다 (NOT_CONFIGURED)"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "summary": "실패한 배포를 AI 로 진단 시작",
        "tags": [
          "diagnosis"
        ]
      }
    },
    "/api/v1/services/{service_id}/deployments/{deployment_id}/diagnosis": {
      "get": {
        "description": "진단 상태(RUNNING·SUCCEEDED·FAILED)와 결과를 돌려준다. 진단 시작 뒤 폴링에 쓴다.",
        "operationId": "get_deployment_diagnosis_api_v1_services__service_id__deployments__deployment_id__diagnosis_get",
        "parameters": [
          {
            "in": "path",
            "name": "service_id",
            "required": true,
            "schema": {
              "title": "Service Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "deployment_id",
            "required": true,
            "schema": {
              "title": "Deployment Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_DiagnosisResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "로그인이 필요하다. 웹훅은 서명이 올바르지 않을 때 (UNAUTHORIZED)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "대상을 찾을 수 없다 (NOT_FOUND · PROJECT_NOT_FOUND · SERVICE_NOT_FOUND · DEPLOYMENT_REQUEST_NOT_FOUND · VARIABLE_NOT_FOUND · DIAGNOSIS_NOT_FOUND · UPLOAD_NOT_FOUND)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "입력이 올바르지 않다 (VALIDATION_ERROR · INVALID_INPUT)"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "summary": "배포의 가장 최근 AI 진단 조회",
        "tags": [
          "diagnosis"
        ]
      }
    },
...
    "/api/v1/services/{service_id}/deployments/{deployment_id}/repair-context": {
      "get": {
        "description": "Trusted repair coordinators receive the original diagnosis-result.v3 and a short-lived source snapshot URL. Select diagnosisId to avoid latest-result races. Treat the URL as a credential; never persist or log it.",
        "operationId": "get_repair_context_api_v1_services__service_id__deployments__deployment_id__repair_context_get",
        "parameters": [
          {
            "in": "path",
            "name": "service_id",
            "required": true,
            "schema": {
              "title": "Service Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "deployment_id",
            "required": true,
            "schema": {
              "title": "Deployment Id",
              "type": "integer"
            }
          },
          {
            "description": "Exact successful diagnosis ID",
            "in": "query",
            "name": "diagnosisId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "exclusiveMinimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Exact successful diagnosis ID",
              "title": "Diagnosisid"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_dict_str__Any__"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "로그인이 필요하다. 웹훅은 서명이 올바르지 않을 때 (UNAUTHORIZED)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "대상을 찾을 수 없다 (NOT_FOUND · PROJECT_NOT_FOUND · SERVICE_NOT_FOUND · DEPLOYMENT_REQUEST_NOT_FOUND · VARIABLE_NOT_FOUND · DIAGNOSIS_NOT_FOUND)"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "현재 상태와 충돌한다. 이미 있는 이름이거나 진행 중인 배포가 있다 (CONFLICT · PROJECT_NAME_CONFLICT · SERVICE_NAME_CONFLICT · VARIABLE_CONFLICT · DEPLOYMENT_IN_PROGRESS · DEPLOYMENT_NOT_FAILED · DIAGNOSIS_IN_PROGRESS · DIAGNOSIS_LOGS_UNAVAILABLE)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "입력이 올바르지 않다 (VALIDATION_ERROR · INVALID_INPUT)"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "외부 시스템(GitHub·Loki·Prometheus·진단 에이전트) 호출에 실패했다 (EXTERNAL_ERROR)"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_NoneType_"
                }
              }
            },
            "description": "필요한 설정이 없다 (NOT_CONFIGURED)"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          },
          {
            "HTTPBearer": []
          }
        ],
        "summary": "소유한 실패 배포의 원본 진단과 단기 소스 URL 조회",
        "tags": [
          "diagnosis"
        ]
      }
    },
...
  },
...
}